add: IDE Screen and update APKStudioScreen
